+++ CHANGELOG.md
@@ -3,9 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.2.0] - 2017-06-08
 ### Added
-- ...
+- Added `symbols` to output options (`data/chars/symbols.js`).
+- Added default values to module (`data/defaults/index.js`). Default values used in cases where `stringScrambler` is invoked *without* an `options` argument.
+
+### Changed
+- Updated `stringScrambler` to support an optional `options` argument.
+- Added `options`-related tests.
+- Moved all output options into `data/chars/` directory. Renamed `chars.js` to `alpha.js`.
+
+## [0.1.1] - 2017-06-07
+### Added
+- Added [Travis CI](https://travis-ci.org/) configuration file to project.
 
 ### Changed
 - Dropped [spread operator](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Operators/Spread_operator) and implemented alternative solution for internal data concatenation.